+++ README.md
@@ -5,21 +5,22 @@ It is a required module of [openimis-be_py](https://github.com/openimis/openimis
 [![License: AGPL v3](https://img.shields.io/badge/License-AGPL%20v3-blue.svg)](https://www.gnu.org/licenses/agpl-3.0)
 
 ## ORM mapping:
-* 
+Registry
 
 ## Listened Django Signals
 None
 
 ## Services
+N/A
 
 ## GraphQL Queries
-None
+N/A
 
 ## GraphQL Mutations - each mutation emits default signals and return standard error lists (cfr. openimis-be-core_py)
-None
+N/A
 
 ## Configuration options (can be changed via core.ModuleConfiguration)
-None
+N/A
 
 ## openIMIS Modules Dependencies
 * core.models.UUIDModel
