+++ govstack_api/_exampleRegistry/registry_config.json
