+++ govstack_api/building_blocks/bb_digital_registries/registries/registry_factory.py
@@ -1,4 +1,7 @@
-from govstack_api.models import Registry
+from typing import Type
+
+from govstack_api.building_blocks.bb_digital_registries.models import Registry
+from govstack_api.building_blocks.bb_digital_registries.registries.base_registry import RegistryType
 from govstack_api.building_blocks.bb_digital_registries.registries.insuree_registry import InsureeRegistry
 
 
@@ -11,9 +14,11 @@ class RegistryFactory:
         if registry is None:
             raise ValueError(f"No registry found for name {registry_name} and version {version_number}")
         registry_class_name = registry.class_name
-        registry_class = globals().get(registry_class_name)
+        registry_class: Type[RegistryType] = globals().get(registry_class_name)
+
         if registry_class is None:
             raise ValueError(f"No registry class found for {registry_class_name}")
+
         registry_instance = registry_class(
             {
                 "class": registry.class_name,
@@ -22,7 +27,9 @@ class RegistryFactory:
                 "special_fields": registry.special_fields,
                 "default_values": registry.default_values,
                 "mutations": registry.mutations,
-                "queries": registry.queries
+                "queries": registry.queries,
+                "registry_name": registry_name,
+                "version_number": version_number
             },
             request
         )
