+++ govstack_api/building_blocks/bb_digital_registries/urls.py
@@ -1,9 +1,12 @@
 from django.urls import path
+
 from .views import (
     SearchRecordView, SingleRecordAPI,
     CheckRecordPresenceView, UpdateOrCreateRecordView,
     MultipleRecordAPI, PersonalDataAPI
 )
+from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView, SpectacularRedocView
+
 
 urlpatterns = [
     path('data/<str:registryname>/<str:versionnumber>/create', SingleRecordAPI.as_view(), name="create_registry_record"),
@@ -17,4 +20,7 @@ urlpatterns = [
     path('data/<str:registryname>/<str:versionnumber>', MultipleRecordAPI.as_view(), name="list_registry_records"),
     path('data/<str:registryname>/<str:versionnumber>/update-entries', MultipleRecordAPI.as_view(), name="update_registry_records"),
     path('data/myPersonalDataUsage/1.0', PersonalDataAPI.as_view(), name="get_personal_data"),
+    path('docs/', SpectacularAPIView.as_view(), name='docs'),
+    path('docs/swagger/', SpectacularSwaggerView.as_view(url_name='docs'), name='swagger-ui'),
+    path('docs/redoc/', SpectacularRedocView.as_view(url_name='docs'), name='redoc'),
 ]
